CefBrowser示例程序：代码结构优化，将Windows平台的代码分离出来，放在Windows目录。
