+++ examples/CefBrowser/Windows/dragdrop/HBitmapControl.cpp
