+++ examples/CefBrowser/Windows/dragdrop/HBitmapControl.h
