+++ examples/CefBrowser/Windows/taskbar/TaskbarManager.cpp
