+++ examples/CefBrowser/Windows/taskbar/TaskbarManager.h
